Added BaseFragment and made changes in NotesFragment, ChecklistFragment, CalendarFragment
