+++ app/src/main/java/com/example/notesvsshoppinglist/ui/base/BaseFragment.kt
@@ -1,0 +1,30 @@
+package com.example.notesvsshoppinglist.ui.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.viewbinding.ViewBinding
+
+open class BaseFragment<VB : ViewBinding>(
+    private val bindingInflater: (layoutInflater: LayoutInflater) -> VB
+) : Fragment() {
+
+    private var _binding: VB? = null
+    protected val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = bindingInflater.invoke(inflater)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
