+++ app/src/main/java/com/example/notesvsshoppinglist/ui/calendar/CalendarFragment.kt
@@ -1,37 +1,21 @@
 package com.example.notesvsshoppinglist.ui.calendar
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.fragment.app.Fragment
 import com.example.notesvsshoppinglist.databinding.FragmentCalendarBinding
+import com.example.notesvsshoppinglist.ui.base.BaseFragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class CalendarFragment : Fragment() {
-
-    private var _binding: FragmentCalendarBinding? = null
-    private val binding get() = _binding!!
+class CalendarFragment : BaseFragment<FragmentCalendarBinding>(FragmentCalendarBinding::inflate) {
 
     private val calendarViewModel: CalendarViewModel by viewModel()
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentCalendarBinding.inflate(inflater, container, false)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
 
-        val textView: TextView = binding.textCalendar
         calendarViewModel.text.observe(viewLifecycleOwner) {
-            textView.text = it
+            binding.textCalendar.text = it
         }
-        return binding.root
     }
 
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
 }
