+++ app/src/main/java/com/example/notesvsshoppinglist/ui/notes/NotesFragment.kt
@@ -1,21 +1,16 @@
 package com.example.notesvsshoppinglist.ui.notes
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import androidx.core.os.bundleOf
-import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import com.example.notesvsshoppinglist.R
 import com.example.notesvsshoppinglist.core.utils.toFormatString
 import com.example.notesvsshoppinglist.databinding.FragmentNotesBinding
+import com.example.notesvsshoppinglist.ui.base.BaseFragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class NotesFragment : Fragment() {
-
-    private var _binding: FragmentNotesBinding? = null
-    private val binding get() = _binding!!
+class NotesFragment : BaseFragment<FragmentNotesBinding>(FragmentNotesBinding::inflate) {
 
     private val notesViewModel: NotesViewModel by viewModel()
 
@@ -32,15 +27,6 @@ class NotesFragment : Fragment() {
         }
     }
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentNotesBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
@@ -49,11 +35,6 @@ class NotesFragment : Fragment() {
         notesViewModel.notes.observe(viewLifecycleOwner) { notes ->
             notesAdapter.submitList(notes)
         }
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
     }
 
     companion object {
