+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -9,8 +9,8 @@ object Config {
     val JAVA_VERSION = JavaVersion.VERSION_1_8
     const val TEST_INSTRUMENTATION_RUNNER = "androidx.test.runner.AndroidJUnitRunner"
 
-    const val VERSION_CODE = 8
-    const val VERSION_NAME = "1.0.7"
+    const val VERSION_CODE = 9
+    const val VERSION_NAME = "1.0.8"
 }
 
 object Modules {
